Revert "Merge branch 'main' of https://github.com/Kitsai/The-Baker-Crow"
This reverts commit 3e8703b6f7e25173f178853fc493029c5a4ab858, reversing
changes made to f8a8a88d0780634b57cc977e880a69f0750c0e9d.

Revert "fix: removing warnings"

This reverts commit ecd399dacd374aae59f3eb173884b75de459ac96.

Revert "fix: Duplicated declaration ID on Update's switch"

This reverts commit 06393e639b202addcc3a6f1bede787fb34c89cb5.

Revert "feat: compile_commands adds"

This reverts commit 6bbe6d0ff795d3c886bb6c53869e5a30617dd03e.

+++ src/Sprite.cpp
@@ -1,7 +1,7 @@
-#include "Sprite.h"
+#include <Sprite.h>
 
-#include "Game.h"
-#include "Resources.h"
+#include <Game.h>
+#include <Resources.h>
 
 #include <iostream>
 #include <algorithm>
@@ -32,7 +32,7 @@ void Sprite::SetFrame(int frame) {
 void Sprite::SetFrameCount(int frame) {
     frameCount = frame;
     currentFrame = 0;
-    associated.box.w = ((double)width)/frameCount; 
+    associated.box.w = width/frameCount; 
 }
 
 void Sprite::SetFrameTime(float frameTime) {
@@ -52,7 +52,7 @@ void Sprite::Open(std::string file) {
         exit(-1);
     }
     SetClip(0,0,width/frameCount,height);
-    associated.box.w = ((double)width)/frameCount;
+    associated.box.w = width/frameCount;
     associated.box.h = height;
 }
